fixes to tests
